+++ src/main/java/kodi/tmdb/TmDbApi.java
@@ -1,10 +1,13 @@
 package kodi.tmdb;
 
 import lombok.val;
+import okhttp3.Callback;
 import okhttp3.Dispatcher;
 import okhttp3.OkHttpClient;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
+import utils.http.OkHttpClientFactory;
+import utils.http.RequestCallback;
 
 import java.util.concurrent.ExecutorService;
 
@@ -12,18 +15,16 @@ public class TmDbApi {
     private final TmDbClient tmDbClient;
     private final OkHttpClient okHttpClient;
     private final String accessToken;
-    private String token;
-    private boolean loggedIn;
 
     public TmDbApi(String accessToken, ExecutorService executorService) {
         this.accessToken = accessToken;
         this.okHttpClient = httpClient(executorService);
-        this.tmDbClient = initClient(accessToken);
+        this.tmDbClient = initClient();
     }
 
+
     private OkHttpClient httpClient(ExecutorService executorService) {
-        val httpClient = new OkHttpClient.Builder();
-        httpClient.setDispatcher$okhttp(new Dispatcher(executorService));
+        val httpClient = OkHttpClientFactory.createOkHttpClient(executorService);
         httpClient.addInterceptor(chain -> {
             val original = chain.request();
             val request = original.newBuilder()
@@ -35,7 +36,7 @@ public class TmDbApi {
         return httpClient.build();
     }
 
-    private TmDbClient initClient(String apiKey) {
+    private TmDbClient initClient() {
         Retrofit retrofit = new Retrofit.Builder()
                 .baseUrl("https://api.themoviedb.org/3/")
                 .addConverterFactory(GsonConverterFactory.create())
